Add admin fe

+++ client-backend/src/main/java/com/dhitha/lms/clientbackend/controller/admin/ClientBookAdminController.java
@@ -1,8 +1,11 @@
 package com.dhitha.lms.clientbackend.controller.admin;
 
 import com.dhitha.lms.clientbackend.client.BookClient;
+import com.dhitha.lms.clientbackend.client.InventoryClient;
 import com.dhitha.lms.clientbackend.dto.BookDTO;
 import java.net.URI;
+import java.util.List;
+import java.util.Objects;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
@@ -13,6 +16,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -26,6 +30,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 @RequiredArgsConstructor
 public class ClientBookAdminController {
   private final BookClient bookClient;
+  private final InventoryClient inventoryClient;
 
   @PostMapping(
       consumes = MediaType.APPLICATION_JSON_VALUE,
@@ -49,8 +54,13 @@ public class ClientBookAdminController {
   }
 
   @DeleteMapping(value = "/{id}")
-  public ResponseEntity<Void> delete(@PathVariable Long id) {
-    bookClient.delete(id);
+  public ResponseEntity<Void> delete(
+      @PathVariable Long id,
+      @RequestParam(required = false, name = "bookReference") List<String> bookReferenceList) {
+    inventoryClient.delete(id, bookReferenceList);
+    if (Objects.isNull(bookReferenceList) || bookReferenceList.isEmpty()) {
+      bookClient.delete(id);
+    }
     return ResponseEntity.noContent().build();
   }
 }
