+++ inventory/src/main/java/com/dhitha/lms/inventory/service/InventoryServiceImpl.java
@@ -2,19 +2,22 @@ package com.dhitha.lms.inventory.service;
 
 import com.dhitha.lms.inventory.dto.InventoryDTO;
 import com.dhitha.lms.inventory.entity.Inventory;
-import com.dhitha.lms.inventory.entity.InventoryId;
 import com.dhitha.lms.inventory.error.GenericException;
 import com.dhitha.lms.inventory.error.InventoryNotFoundException;
 import com.dhitha.lms.inventory.repository.InventoryRepository;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
 
 /**
  * Implementation for {@link InventoryService}
@@ -22,6 +25,7 @@ import org.springframework.stereotype.Service;
  * @author Dhiraj
  */
 @Service
+@Slf4j
 @RequiredArgsConstructor
 @Transactional
 public class InventoryServiceImpl implements InventoryService {
@@ -59,38 +63,60 @@ public class InventoryServiceImpl implements InventoryService {
 
   @Override
   public long getAvailableCount(Long bookId) {
-    return inventoryRepository.countByIdBookIdAndAvailable(bookId,true);
+    return inventoryRepository.countByIdBookIdAndAvailable(bookId, true);
   }
 
   @Override
-  public void add(InventoryDTO inventoryDTO) throws GenericException {
-    try {
-      inventoryRepository.saveAndFlush(mapToEntity(inventoryDTO));
-    } catch (DataIntegrityViolationException e) {
-      throw new GenericException(
-          String.format(
-              "Book with id %s already has reference id %s",
-              inventoryDTO.getBookId(), inventoryDTO.getBookReferenceId()),
-          HttpStatus.BAD_REQUEST.value());
+  @Transactional(rollbackOn = Exception.class)
+  public List<InventoryDTO> add(InventoryDTO inventoryDTO, Integer count) throws GenericException {
+    Assert.notNull(inventoryDTO, "Inventory cannot be null");
+    count = count != null && count > 0 ? count : 1;
+    Set<Inventory> set = new HashSet<>(count);
+    while (count-- > 0) {
+      Inventory inventory = this.mapToEntity(inventoryDTO);
+      inventory.getId().setBookReferenceId(UUID.randomUUID().toString());
+      set.add(inventory);
     }
+    try {
+      inventoryRepository.saveAll(set);
+    } catch (Exception e) {
+      throw new GenericException(
+          "Error while saving inventory", HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
+    return set.stream().map(this::mapToDTO).collect(Collectors.toList());
   }
 
   @Override
   public void delete(Long bookId) throws InventoryNotFoundException {
-    try {
-      inventoryRepository.deleteByIdBookId(bookId);
-    } catch (EmptyResultDataAccessException e) {
-      throw new InventoryNotFoundException("No inventory found for book id: " + bookId);
+    List<Inventory> bookInventory = inventoryRepository.findByIdBookId(bookId);
+    if (bookInventory.isEmpty()) {
+      return;
     }
+    long count = bookInventory.stream().filter(inventory -> !inventory.getAvailable()).count();
+    if (count > 0) {
+      throw new InventoryNotFoundException(
+          "Some books of book id " + bookId + " are loaned and cannot be deleted ");
+    }
+    inventoryRepository.deleteByIdBookId(bookId);
   }
 
   @Override
-  public void delete(Long bookId, String bookReferenceId) throws InventoryNotFoundException {
-    try {
-      inventoryRepository.deleteById(new InventoryId(bookId, bookReferenceId));
-    } catch (EmptyResultDataAccessException e) {
-      throw new InventoryNotFoundException(
-          String.format("No Book found with id %s and reference id %s", bookId, bookReferenceId));
+  @Transactional(rollbackOn = InventoryNotFoundException.class)
+  public void delete(Long bookId, List<String> bookReferenceIdList) throws InventoryNotFoundException {
+    for(String bookReferenceId: bookReferenceIdList) {
+      Optional<Inventory> inventory =
+          inventoryRepository.findByIdBookIdAndIdBookReferenceId(bookId, bookReferenceId);
+      if (inventory.isEmpty()) {
+        throw new InventoryNotFoundException(
+            String.format("No Book found with id %s and reference id %s", bookId, bookReferenceId));
+      }
+      if (!inventory.get().getAvailable()) {
+        throw new InventoryNotFoundException(
+            String.format(
+                "Book with id %s and reference id %s is loaned and cannot be deleted",
+                bookId, bookReferenceId));
+      }
+      inventoryRepository.deleteByIdBookIdAndIdBookReferenceId(bookId, bookReferenceId);
     }
   }
